backend/requirements.txt
celery==5.0.2
Django==3.1.3
django-cors-headers==3.5.0
django-health-check==3.16.1
django-redis==4.12.1
psycopg2-binary==2.8.6
djangorestframework==3.12.2
django-fsm==2.7.1
django-nose==1.4.7
coverage==5.3
drf-yasg==1.20.0
djangorestframework-api-key==2.0.0
django-expiring-token==1.0.2
reportlab==3.5.55
stripe==2.55.0